Move ConfirmationModal to components/common
