+++ airbyte-webapp/src/components/common/ConfirmationModal/ConfirmationModal.module.scss
