+++ airbyte-webapp/src/components/common/ConfirmationModal/index.ts
